Added initial login page and initial user model
User model is not set in stone, but set up an initial model.

Also added the login page draft, as discussed. Though it has no logic, but the buttons and form fields are set up to be added logic to. Added route to the new page on the login or sign up text in drawer, added inkwell to it, to make it interactive.