Added Gemini parsing
Set up a model for the recipe, and used chatgpt to help me with the regex for the parsing via factory method of the recipe model class.

Also set up the recipe screen to be dynamic with the recipe data, still need to store recipes in Supabase, and use that instead. But this is a step.