Removed Prints (lint errors), moved icons
To keep all navigation logic more streamlined, we decided to move the drawer icon to the bottom left side, instead of top left. Made some alterations to the router file then, to make sure the highlighting and such works as intended. Still need to set up the topbar of the screens individually, and insert the searchwidget.