Fixed system back navigation
Still need further testing to make sure all is working as expected.

However, changed the logic around to use pushNamed instead, so that it adds to the stack, and the system button automatically searches for stack, and the previous go did not work for this.

Set up an AppShell helper file as well for keeping track of how the main layout of the tabs should be.

Back button now works if there is a valid screen to return to!