Allows user to save recipe from recipe screen
Added a heart icon now next to the recipe name, which allows the user to save the recipe to supabase, though this method needs to be refactored and probably pulled more into service / controller :).

Also allows each ingredient to be clickable, and added a button (with alterations) that uses for add all.