Added Purchase Statistics screen
Added a static screen with dummy data for displaying a pie chart of money spent across various categories in a month.

Also set up a line chart for displaying money spent over the year(s) with an average in a line, where it shows the months where user spends above or below the average.