Temporary refactor to fix extremly slow screen
The app was slow and rebuilding alot because each shippinglistitem contained a dropdown that could list up to 5000 quantites. Since we displayed this dropdown for every item in the list the framework had to build and rebuild all those dropdowns entries. Caused a huge performance overhead.

Replaced large dropdown with a simple text field to edit quantity for now.