Prevented potential race condition
As the service had not been updated yet to use the new geminiRaw with the edge function, and expected the gemini flutter package public key, there was a race condition on just my phone, but not others. Fixed, as I was the one with problem, and could actually test and implement fix.