Fixed meal suggestions + snackbar
As I had mistakenly used the wrong database table, the logic for meal suggestions was NOT working as intended. Have updated it to use the atual pantry (inventory) now, so that when user removes something, it doesn't constantly get suggestions based upon that.

Additionally, have added the custom snackbar to the ingredients list (meal screen).